refactoring. moved restapi access point to main.py and removed old files
